creates a database abstraction (postgresql/sqlite)
Used sqlite for dev environment.
Used postgresql for prod environment.

Correction of types for entities, so that it is compatible between both databases.

+++ src/contracts/contract_shared/responsible.v
@@ -1,6 +1,6 @@
 module contract_shared
 
-pub enum Responsible as i8 {
+pub enum Responsible as int {
 	pai = 0
 	mae = 1
 }
